requirements.txt
requests
beautifulsoup4
selenium
schedule
python-dotenv